Simplified the core implementation of B* and changed its interface with binder macros, making them more flexible and forcing them to do more of the work.  In particular, there are no more special suffixes such as + and @ for binder names that change the way the core deals with them.  For every binder FOO,
(b* (((foo a b c) (asdf x y))) bla)

expands to

(patbind-FOO (a b c) ((asdf x y)) bla)

This means that the definition of the macro PATBIND-FOO may need to do
more work, such as binding (asdf x y) to a variable so that it won't be
evaluated more than once; but it has the flexibility to not do that if
that is appropriate, which was a problem in the old design.

Finally, symbol packages are now respected when constructing the
patbind- macro calls.  Whatever FOO's package is above is also the
package of PATBIND-FOO.