Quoting from :doc note-4-2:
Sped up ~ilc[include-book] a bit by memoizing checksums of symbols.  (This
change pertains to ``normal'' ACL2 only, not the ~ilc[hons] version
(~pl[hons-and-memoization], where such memoization already occurred.)  We
found about a 23% speed-up on an example from Dave Greve.

Fixed a bug that was failing to substitute fully after binding free variables
in ~ilc[force]d hypotheses.  Note that the binding of such a free variable
had been to a new variable named something like ~c[???-Y], but the name is
now of the form ~c[UNBOUND-FREE-Y].

Enhanced accumulated-persistence utility to track hypotheses and
conclusion of a rune if suitable arguments are given to
accumulated-persistence and show-accumulated-persistence.