Arranged that local events are no longer elided when writing out a .acl2x file (bug reported by Sol).
Close output channel when writing .acl2x file (bug reported by Sol).

Fixed a bug in which the wrong attachment could be made when the same
function has an attachment in a book and another in the certification
world of that book (possibly even built into ACL2), if the load of a
compiled file is aborted because a sub-book's compiled file is
missing.