Quoting :doc note-4-3:
Calls of mbe in "safe-mode" situations -- i.e., during evaluation of
defconst, value-triple, and defpkg forms, and during macroexpansion --
are now guard-checked.  Thus, in these situations both the :logic and
:exec forms will be evaluated, with an error if the results are not
equal.  Formerly, only the :logic form was evaluated, which was a
soundness bug that could be exploited to prove nil.  For a such a proof
and a bit of further explanation, see the example at the top of the
comments for (deflabel note-4-3 ..) in ACL2 source file ld.lisp.