(GCL only) Fixed a bug in set-debugger-enable.
Slightly improved the message that is printed just below a hard Lisp
error.  Thanks to Francisco Palomo Lozano for a suggestion that led to
this change.

(Lispworks mods courtesy of Martin Simmons via David Rager) Changed
calls of special-form-or-op-p to expand to calls of
special-operator-p.  Changed 'cl::getenv to 'hcl::getenv and changed
'cl::setenv to 'hcl::setenv.