Made small comment and :doc improvements, including ACL2(p) stuff in :doc note-4-4 (in consultation with David Rager).
Added comment in memoize-fn (in consultation with Sol Swords) about
space taken by memoization with :forget t.

Added "for ACL2(p): " prefix for one-liners in :doc topics in section
SWITCHES-PARAMETERS-AND-MODES that pertain to ACL2(p).

Fixed (I hope!) ev-fncall-rec-logical to comprehend insistence on
guard-checking for primitive stobj functions that take the latched
stobj as an input.

We now print notes to use :a! and see :DOC set-evisc-tuple when evaluating
:brr t.  Thanks to Robert Krug for suggesting this improvement.