- FIxed a limitation where multi-valued functions caused the default   simple strategy of cgen/testing errored out. The way I fix it, is by   gathering all such function calls and wrapping them with (mv-list k call).
- As Matt rightly remarked that Cgen should gracefully skip/ignore
  forms that Cgen cannot support i.e test/execute. Currently constrained
  functions with no attachments and functions with stobjs arguments are
  not supported and Cgen simply skips searching for cts/wts after printing
  a warning message.

Note: There is still a matter of a subgoal being supported by Cgen, but the
top-level goal is not. In this case printing cts/wts might result in
an error, i.e. stobjs restrictions not met. The right way to deal with
unsupported conjectures, is to identify it at the top/event level itself,
rather than checking at subgoal/checkpoint level. But since right now
we dont have access to form/body in initialize-event-user, I am leaving
this known bug (in Cgen) as it is for the moment.