Add new std/alists directory.  For now this is pretty minimal.
Replace some of vl/util/arith.lisp with the new std/alists stuff.

Add cutil/defruled, which is to defrule as defthmd is to defthm.

Localized a couple of general-purpose alist lemmas in witness-cp, since they
aren't really what the book is about.

Slight revisions of vl-lint, adding some minor features and documentation.