Restored handling of leading "~" in pathnames on Windows, but retained the fix that avoids an error on Windows, except from inside certify-book.  Here is the revised item from :doc note-6-2, which has a new sentence at the end.
| (Windows only) Fixed a bug that was causing a hard error on Windows when ACL2
| encountered filenames starting with the tilde character (~), for
| example, (ld "~/acl2-customization.lsp").  Thanks to Sol Swords for
| bringing this bug to our attention.  Also thanks to Harsh Raju Chamarthi for
| a useful conversation that led to a better fix than our first one.

Added a comment about prohibiting verify-termination for a
conditionally memoized function.

Added a link from print-summary-user to its replacement,
finalize-event-user.

Replaced a couple of calls of our-with-standard-io-syntax by
with-standard-io-syntax in the memoization code, since ANSI GCL now
seems to handle with-standard-io-syntax.  Accordingly modified
our-with-standard-io-syntax.

Modified get-output-stream-string$-fn so that the special handling of
types for GCL is now restricted to non-ANSI GCL, since ANSI GCL now
gets that right.

Added targets regression-everything and regression-everything-fresh in
GNUmakefile.

Cause an error in ANSI Lisps when COMMON-LISP::*PRINT-PPRINT-DISPATCH*
is not bound, since this has been fixed for ANSI GCL.

Finally, quoting a comment in (deflabel note-6-2 ...):

; For GCL only, further increased the binding stack (except, still, on
; Windows); see the setting of si::*multiply-stacks* in init.lisp.