Made miscellaneous comment and documentation mods.
Deleted deprecated GNUmakefile targets regression-legacy-fast and
regression-legacy-fast-fresh.

Added a comment to note-6-2 about something already done before the
release: "For SBCL only, increased --control-stack-size from 4 to 8."

Increased the stack size for CMUCL, as per the following comment
in the definition of function save-acl2-in-cmulisp-aux.

; In our testing for ACL2 Version_6.2 we found that certification failed for
; ACL2(h) built on CMUCL for the book tau/bounders/elementary-bounders.lisp,
; with the error: "CMUCL has run out of dynamic heap space (512 MB)."  This
; failure doesn't seem to be fully reproduceable, but it seems safest to
; increase the stack size.  Our CMUCL image, even though on 64-bit linux,
; reported the following when we tried a value of 2000 here:

; -dynamic-space-size must be no greater than 1632 MBytes.

; So we use 1600 and see how it goes....

Thanks to a suggestion from Jared, deleted obsolete name mfc-ap-ttree
from *primitive-logic-fns-with-raw-code* and wrote code in
check-built-in-constants that checks for such names.  But I left that
code as a comment, since it's not yet clear how to do such testing in
a way that works not only with ACL2 but with all variants of ACL2
(e.g. ACL2(p)).