Made the following changes (quoting comments in (deflabel note-6-3 ...)):
; Improved the error message for #.expr when expr hasn't been defined by
; defconst.

; Modified printing of the banner for GCL, especially for versions 2.6.9 and
; later, due to a change in how GCL prints such banners.  (Printing of the
; banner still works fine with older versions of GCL.)

The rest are comment changes:

Updated installation/misc.html to point to new acl2-books Google
groups list.  (I might change this again after talking with J.)

Improved comment for trans-or.

Added comments to expand-and-or that explain why its calls of
expand-abbreviations are not redundant.

Added a comment in *definition-minimal-theory* explaining the
inclusion of mv-nth.

Improved comments in make-certificate-file-relocated to clarify why we
don't replace strings (for creating relocated .cert files for Debian
GCL distributions).