centaur/gl/solutions.lisp: Made an adjustment so that example could be handled by CMUCL.
xdoc/mkdir-raw.lisp:
Avoided problem making the manual using Lisps that don't compile on
the fly, such as LispWorks.  An error was caused, with xdoc::mkdir
printing "raw lisp definition not yet installed", I think because the
progn! redefinition of that function using raw Lisp was creating an
interpreted definition that was getting replaced, using the ACL2
definition, by (comp t).  I solved the problem as suggested in
:doc comp, by adding to state global 'logic-fns-with-raw-code.

misc/bash-bsd.lisp,
misc/bash.lisp:
Added :xdoc documentation, and put a BSD license on the code, which is
in bash-bsd.lisp.  As explained in a note near the top of
bash-bsd.lisp, I hope that the documentation will also be licensed
that way, but for now I've played it safe and put a GPL v2 license on
it.