Avoid accidental, futile attempt to certify rendered-doc.lisp, by calling it rendered-doc.lsp instead.
Updated documentation for NOTE-6-4 and DOCUMENTATION.  The latter may
be improved in the future.

Modified system/doc/render-doc.lisp so that its output file,
system/doc/rendered-doc.lsp, would be printed in the ACL2 package.
(Avoiding package prefixes can be helpful for writing Emacs tools to
browse that output file.)