Unify ACL2(r) books building with the ordinary books build
-------------------------------------------------------------------------------

Aside from one small issue related to proof outlines for the 1999 workshop
calculus books, I think this commit nicely integrates the ACL2(r) build into
books/GNUmakefile, eliminating the need for books/nonstd/Makefile.

It would be great if Matt and Ruben can take a look at this.  Once everyone is
happy, we can merge it into trunk.

The rest of this message explains what I did, how to work with the new system,
and what remains to be done.

-------------------------------------------------------------------------------

Part 1: Basic Steps

I started by extending cert.pl with new parameters, so that individual books
can say whether they require or are incompatible with ACL2(r) using cert_param
comments, e.g.,:

     ; cert_param: (uses-acl2r)
     ; cert_param: (non-acl2r)

Then I extended GNUmakefile so that:

  - if we're using ACL2(r), then it will not try to build any books that
    are marked as (non-acl2)

  - if we're not using ACL2(r), then it will not try to build any books
    that are marked as (uses-acl2r)

Then just started trying to build the regular books with ACL2(r), e.g.,:

     $ cd acl2-sources/books
     $ make ACL2=acl2r -j 4

This of course didn't work, because there are many books that don't certify
with ACL2(r).  For instance, various low-level RTL books failed when trying to
prove something about rationalp or realp.  Some problems were weirder.
Defsort/remove-dups failed because remove-adjacent-duplicates is apparently
built into ACL2(r) but not into ACL2.  Other books failed due to
CHK-ACCEPTABLE-DEFUNS having a different number of arguments.

I never tried to fix any of the failures.  Instead, to work around them, I
simply marked any offending book with (non-acl2r).  The effect of this is to
tell Make to not certify these books or (transitively) any books that include
them when ACL2(r) is being used.  In the end, there are many, many books that
get transitively killed off.  But I think that's just fine, and I don't know
how to do better without actually trying to fix up the failures.

At any rate, for the most part the above was sufficient.  But in a couple of
cases I had to do more than add these non-acl2r comments.  The most major case
was in the arithmetic/ library.  Here, a couple of include-books were wrapped
in #+non-standard-analysis, and appeared to exist only in the nonstd/arithmetic
directory.  And this basically doesn't work at all.

In order to use the same build for both ACL2 and ACL2(r), I think we need to
make sure that the include-books are always the same, regardless of which
variant is being used.  Fixing this seemed pretty easy.  I just:

   - moved these books into from books/nonstd/arithmetic into books/arithmetic
   - changing arithmetic/top to unconditionally include them, and
   - changed the contents of real.lisp, etc., to be, e.g.,:

       #+non-standard-analysis
       (progn ... everything that was previously in the book ...)

This seems like a good solution.  It keeps the include-books the same, and lets
the books coexist in a pretty straightforward way.  I think the only other book
that had this issue was workshops/2004/gameiro-manolios/support/top-with-meta,
but this was easily fixed by just removing the conditionals from the includes,
and into the included books themselves.

Anyway, I added these annotations to every book that failed until "make all"
finished successfully using ACL2(r).  (I didn't try a "make everything" because
I'm on my home computer.)

-------------------------------------------------------------------------------

Part 2: Getting nonstd/* working with ACL2(r)

The next step was to get books/GNUmakefile to also build the books underneath
nonstd/ when ACL2(r) was being used.

Most of this was very straightforward.  I removed "nonstd" from the grep
command in GNUmakefile that excludes the nonstd books and tried to build.  To
get things working, I had to make some minor adjustments to include-books
because, previously, it appears that the books in nonstd were expecting :dir
:system to be configured to point at books/nonstd.  But now we are trying to
certify them while :dir :system points at books/.

Similarly, I ended up getting rid of books/nonstd/Makefile-generic and just
re-working the existing Makefiles that used it to point to
books/Makefile-generic.  (I was able to delete most of these eventually.)

Anyway, almost all of this was very smooth.  The only remaining problem is that
in the workshops/1999/calculus directory, which has a generated book
(tree.lisp) and also some kind of proof-outline stuff which I know nothing
about, but which appears to need some custom Makefile commands.  By porting the
commands from books/nonstd/Makefile into books/Makefile, this is working fine
at least insofar as certifying the generated tree.lisp file.  However, the
proof-outline stuff isn't working for some mysterious reason.  It is somehow
not opening up the "nonstd/nsa/nsa.lisp" file.  I haven't spent a lot of time
trying to figure this out, but maybe someone who knows more about this can take
a look at, and figure out a way to fix it.

At this point, the following command appears to build all of the nonstd books
and all of the regular books that I didn't mark with (non-acl2r), and their
dependents:

    $ cd acl2-sources/books
    $ make -j 4 all ACL2=acl2r

-------------------------------------------------------------------------------

Part 3: Getting nonstd/* working with ACL2(non-r)

This part was really easy.  I just started trying to build "make all" using
ordinary ACL2, and added

   ; cert_param: (uses-acl2r)

To the books that failed.  It didn't take many annotations to transitively
kill whatever didn't work.

This is a nice solution and is, I think, superior to just excluding nonstd/*
from the build.  Something really neat about it is that, now, there's not
really anything special at all about the nonstd/ directory.  If we want to
move, e.g., the nonstd/workshops books into the ordinary workshops/ directory,
it should be quite easy to do that (just move the files and fix up whatever
include-books break.)

-------------------------------------------------------------------------------

Anyway, I think that's it?  Matt will probably want to update some things in
acl2/GNUmakefile, etc., but otherwise this seems good?