Updating boothmul to use the newer version of stv-decomp-proofs.
Also adding an example use of specific inputs and hyps to show how
that approach can break.  The tutorial may not be the best place for
that, but I didn't want to create an additional
boothmul-noautoins.lisp

The hint required to make stv-decomp-theory work is now smaller, which
is kind of nice.