Added missing local special declaration of *default-hs* for ACL2(hp).
All I did was to add special declaration for *default-hs* under its
let-binding in consume-work-on-work-queue-when-there.  I agree that as
per CLHS, this shouldn't be necessary, though perhaps I'm missing
something.  (For example, I didn't check whether Lisp is required to
make special variables thread-local, so for all I know, maybe CCL has
a right not do that unless the special declaration is explicit.)

I'm surprised at Jared's comment, "But that doesn't seem to work
either", when he tried something similar.  But anyhow, this fix
seems to work.