Fixed bug pointed out by David Rager, which made SBCL builds fail for ACL2(hp).
I think the problem had to do with recent changes I've made involving
with-lock and (recently added) with-lock-raw.  The fix was to change
calls of with-lock in multi-threading-raw.lisp to with-lock-raw, since
with-lock is defined in axioms.lisp, which is loaded after
multi-threading-raw.lisp.  I'm a little surprised that this bug didn't
show up in CCL.  And, this is another case where I'm glad we support
multiple Lisps.

(Normally I run a full regression, but this change can only affect
ACL2(p) and ACL2(hp) and builds work fine, so checking in.)

Also, added a comment for state global acl2-sources-dir saying why it
should not be deleted (Jared now uses it for XDOC).