Satlink: add optional book for gathering benchmarks
This adds a Satlink plugin for collecting the DIMACS files that Satlink sends
to the SAT solver.

If you have hard GL/Satlink proofs, you might use this tool to gather up the
problems that are being given to the SAT solver.  You could then send the
resulting DIMACS files to the SAT competition folks to use as benchmarks, or
use them to experiment with different SAT solvers to see which ones are the
best on your particular problems.