Minor xdoc tweaks
Arrange so that defxdoc, defsection, and define evaluate the arguments to
:short and :long instead of quoting them.  This may allow you to write
macros that do stuff like:

   `(defsection foo
      :short (cat "blah blah" ,(symbol-name blah-p) "blah blah.")
      ...)

Updated defsection documentation to include a discussion of the recently
added commentary feature, i.e.,

    (defsection foo
      :parents ...
      :short ...
      (defun f (x) ...)
      "<p>The next theorem shows that F is really great.</p>"
      (defthm f-is-the-best ...))

Tweaked defsection so that you can give :extension (foo) instead of just
:extension foo.  (You still can't give :extension (foo bar), but that's
a perfectly sensible restriction.)

Tweaked deflist so that it can create documentation even in the
already-definedp case.

Better error handling on xdoc-extend and xdoc-prepend.

Added a few tests cases.