Fixed a bug in fast-alist-clean found by Sol Swords, and replaced maphash with clrhash.
That bug (use of car on an atom) never seems to have caused problems,
but it was worth fixing.  I believe that I introduced the
fast-alist-clean bug after Version 6.5 (hence no entry in note-7-0 for
this).

Thanks to Sol for finding this, and also for comments (Jared too)
suggesting the replacement of maphash with clrhash in
hl-hspace-fast-alist-clean.  Quoting a new comment there:

; .... In an experiment with the seven
; supported Lisps, only CLISP reduced the size of the hash-table when using
; clrhash.  Otherwise, clrhash is preferred, at least in CCL where maphash
; appears to do some consing while clrhash does not.

So I made this change only for #-clisp.

Unrelated minor change: Tweaked books/system/toothbrush/README just a bit.