Updated toothbrush Makefiles to use TB_LISP instead of LISP and to avoid errors when using certain shells.
Jared Davis reported inadvertent attempts to do toothbrush testing
caused by his (quite reasonably) setting LISP, so I changed that
(thanks, Jared).

The shell issue was reported by David Rager (thanks).

Added role of acl2r.lisp to README.

Extended test2 so that it checks that system functions are
automatically defined as needed.  Thanks to Keshav Kini for asking a
question in the ACL2 seminar that led to this extension.