Made documentation changes, in particular (and mostly) to complete the reorganization of topics under the ACL2 topic that was begun by Jared.
In the process, I dealt with nearly all the xdoc::change-parents calls
in xdoc::fix-the-hierarchy.  Also, I made some decisions that could
well be revisited, such as changing the parents of UNSOUND-EVAL from
MISCELLANEOUS to (INTERFACING-TOOLS PROGRAMMING-WITH-STATE).  In
addition, I changed each defsection to defxdoc in
books/system/doc/acl2-doc.lisp, simply to make it easier to search
(with "defxdoc <topic-name>").