Fixed bug in interaction of :a! with proof-checker and :brr.
  When aborting with :[a!] inside the [break-rewrite] loop from within
  the [proof-checker], a raw Lisp error could subsequently occur
  saying, ``Attempt to execute *wormhole-cleanup-form* twice!''. This
  has been fixed. Here is an example that formerly caused that
  behavior.

    (defthm silly (implies (null x) (equal (append x y) y)))
    (defconst *c* (make-list 10))
    :brr t
    :monitor (:rewrite silly) t
    (verify (equal (append x (append *c* y)) (append y (append *c* x))))
    bash
    :go
    :a!
    bash ; formerly caused error

Technical note: The fix was to re-work the way that one remains in the
proof-checker after a lisp error.

Note: As usual, I ran a regression before checking in, but there was
initially a failure for the book centaur/misc/alist-equiv, as follows:

  Error:  Fault during read of memory address #x0

Subsequent certification attempts succeeded.  I'm pretty confident
that this was just a CCL (Version 1.11-dev-r16288M-trunk) glitch.