Clarified in :doc note-7-0 that ACL2(c) may be eliminated after the next release after 7.0 [hence, presumably, after 7.1].  Improved error message for memoize on a symbol.  Clarified in installation instructions that directory shouldn't contain whitespace.  Added WATCH documentation.
Thanks to Jared Davis for bringing up the idea of eliminating support
for ACL2(c).

A call such as (memoize foo) is an error because memoize evaluates its
argument, and is to be executed at the top level -- for example,
(let ((x 'foo)) (memoize x)) fails.  We have improved the error
message when the first argument of memoize is a symbol, to suggest
explicitly that the argument needs to be quoted.

Thanks to Bob Boyer for pointing out the directory whitespace issue.

Thanks to Bob Boyer for providing most of what is in :doc watch.