Split VL into "VL" and "VL2014", the "2014 Edition" of VL.
VL2014 is a "stable" fork of VL.
  - It lives in a new directory: books/centaur/vl2014
  - It uses the "VL2014" package.
  - We intend for it to continue to work with ESIM and other older tools.
  - We do not intend to actively develop it further.

VL continues as the "main" version of VL.
  - It continues to live in: books/centaur/vl.
  - It continues to use the "VL" package.
  - It continues to be under active development.
  - It no longer supports ESIM.

ESIM has been adjusted to include books from VL2014.

More information about this split, from books/centaur/vl2014/doc.lisp:

-------------

<p>The <u>2014 Edition</u> of VL is a \"stable\" fork of VL that is meant to
provide continuing support for older tools, notably including @(see
acl2::esim), and also including other internal Centaur tools that are based on
earlier versions of VL.</p>

<p>The 2014 Edition is <b>not</b> under active development.  You should
probably not use it for new projects, and it may be best to work toward
eventually transitioning older tools away from it, in favor of the \"main\"
version of @(see vl::vl).</p>

<p>Note that you can tell which version of VL you are using as follows.</p>

<ul>
<li>Main Edition: @('centaur/vl') directory, @('vl') package.</li>
<li>2014 Edition: @('centaur/vl2014') directory, @('vl2014') package.</li>
</ul>

<p>Since each version uses its own package, it is possible to load both
versions into the same ACL2 project simultaneously.  This may be of use when
transitioning code from the 2014 Edition to the main edition.</p>

<h3>Why is there a 2014 Edition?</h3>

<p>VL has a long history.  Its development started in 2008.  At that time, our
goal was only to implement support for Verilog-2005.  Over the next several
years, we used VL to implement many tools, many of which were based around
@(see acl2::esim).  ESIM is a purely bit-level backend.</p>

<p>In 2014, we began extending VL with features from SystemVerilog and also
began work on a new, vector-level backend, @(see svex).  Throughout 2014, and
into the start of 2015, we made substantial changes to VL and transitioned our
internal projects to SVEX.</p>

<p>Much of this work was compatible with ESIM and expanded what the VL/ESIM
flow could handle.  For instance: during this time, the ESIM flow gained better
support for ANSI-style ports, SystemVerilog style global parameters, basic
SystemVerilog idioms like @('always_comb') and @('logic') data types, and so
on.  Other certain SystemVerilog features (structs, arrays, etc.) would have
been difficult to support in ESIM, so we only added them to the VL/SVEX
flow.</p>

<p>In February 2015, we began working on a substantial change to VL's
expression representation.  It became apparent that supporting ESIM through
this change would require updating a large body of code.  This would be a major
cost to maintain code that we were no longer using.</p>

<p>In light of this, we decided that it was time to end support for the ESIM
flow in VL, but we did not want to abandon ESIM altogether.  As a compromise,
we decided to create a stable fork of VL, based on the last version of VL that
works with ESIM.  This fork would allow older tools and proofs to continue to
function, while giving us the freedom to move more quickly on the new versions
of VL without having to continually update the ESIM flow.</p>

------------------