Adjust xdoc jump-to box order.
The jump-to box now prefers exact matches and, subsequently, returns topics in
their rough importance-order instead of in alphabetical order.

I think this probably resolves #344.  It may be that we want to make additional
tweak to the importance ordering or to the jump-to box itself.

I would appreciate feedback on how well this is working for folks, after they
have tried it out for awhile.