Remove ccl-only cert_params from satlink-based books
These used to be marked as ccl-only because they used tshell.  I think they
should now be portable to all Lisps except for GCL, due to lack of Quicklisp
support.  I think we won't need a non-gcl cert_param because the Makefile
"deeply knows" that Quicklisp doesn't work on gcl?  So I'm just going to push
this and hopefully it doesn't break anything.