Fixed time-tracker bug.  Put term-order and merge-sort-term-order in :logic mode.  Fixed home page to match a change previously made on the web.  Fixed acl2-doc bug in "S".
Quoting :doc note-7-2 (which was just added); thanks to J for adding
the guards and doing the proof development for the second item (on
term-order etc.).

  When ACL2 was interrupted while debugging was on (see
  [set-debugger-enable]), it was possible later to get the following
  error repeatedly:

    HARD ACL2 ERROR in TIME-TRACKER:  It is illegal to specify :START for
    tag :TAU, because tracking for this tag is already in an active state.

  This problem has been fixed, by defining a new keyword argument for
  [time-tracker], :start!, and using it to track the use of the
  [tau-system]; see [time-tracker] and [time-tracker-tau].

  The built-in functions term-order and merge-sort-term-order are now
  [guard]-verified in :[logic] mode. (Technical note: This
  enhancement followed the usual process: added a community book,
  books/system/merge-sort-term-order.lisp, included that book in
  books/system/top.lisp, and updated source constant
  *system-verify-guards-alist* accordingly.)

  We fixed a bug in the [ACL2-doc] command, S (command
  acl2-doc-re-search).

Improved comments in the section "Support for system-verify-guards" of
source file boot-strap-pass-2.lisp.

Tweaked "Documenting Your Books" section in :doc documentation.

Fixed home page to match a change previously made on the web for 7.0,
which is now visible on the web for 7.1 as well, to clarify that ACL2
is also a logic and to show what "ACL2" stands for.