Alpha version of SV.
This merges in Centaur's development branch for VL and SV.
(SV is a successor to ESIM.)

VL and SV are still in heavy development and you should not
try to use them yet.  We still have a lot to do in terms of
testing and documentation, and we fully expect that VL and
SV will continue to change rapidly in the near-term.

(This push is mainly intended to allow us to identify any
unforseen problems with it (e.g., maybe it doesn't work on
Allegro or something) and to start more thoroughly testing
SV in our regressions.)