Allow stobjs in GCL with more than 64 fields; two other, minor changes.
Quoting :doc note-7-2:

  (GCL only) It had not been possible to define a [stobj] with more
  than 64 fields in GCL. We have removed that restriction. (Technical
  note: GCL disallows calls of the function, vector, with more than
  64 arguments. So instead we now build a list of stobj fields that
  is coerced to a vector, rather than calling vector directly.)

Changed a conditional ignore declaration in set-gc-strategy-fn
to an unconditional ignorable declaration.

Expanded the discussion of the :ttag argument in
define-trusted-clause-processor, triggered by an acl2-help query from
Yan Peng.

With help from Shilpi, fixed a couple of broken x86isa :doc topics.

Note: The ACL2-Doc Emacs browser is broken because of the (apparently
new) topic, SV::4VEC-[=, which uses the [ character.  I'll try to fix
it tomorrow, but I'm not sure how easy a fix this will be.