Changing defsvtv's autoins-lookup-casesplit function so that it doesn't provide a casesplit hint when there are no cases.
I have some concern that the lack of cases in my very large example might actually mean that something else is amiss, but I'm fairly sure that the problem isn't with the circuit.  If something is amiss, it would more likely be that the autoins-lookup theorem could use some adjustment.  Either way, I think this is a reasonable solution for now, as it allows our massive defsvtv to go through.