turn off the annoying list/cons evalable printing which is enabled by default by the book evalable-ld-printing.
Added first-class support (like for equality) for membership
relation/constraint to Cgen. i.e member/enum has the same status as
range types in Cgen

Added support for collecting stats on which hypotheses are failing in
vacuous tests

cgen/testing which is put into effect using override-hints, now uses
override-hints! i.e. it is not local to books anymore (this was done
to collect regression-wide stats and probably will be retracted later.

dont return IF as a sub-function of a term.

cgen-timeout is now a global timeout value for test?. Change the name
later to a more appropriate one. The call to prove/cgen is now under
the with-prover-time-limit wrapper. The original cgen-timeout has been
renamed cgen-local-timeout and only puts a bound on the testing/search
computation by cgen for each local subgoal. The default value of
cgen-local-timeout is still 10 seconds but the global cgen-timeout
which restricts time spent by the prover in test? is 3600
seconds.

Major simplification: Give up support for nested testable events (like
thm, defthm etc); this results in the simplification of the
event-stack global to just a event-ctx global which stores the
ctx-form and on even/PRE resets it to that events ctx-form and on
event/POST checks whether its ctx-form matches the stored event-ctx;
if no skip, if yes, reset event-ctx. This makes cgen::flush
unnecessary/irrelevalt. So if you interrupt a long running thm and run
another thm, the cgen results from the first will not
pollute/spill-over the second thm form. This also now allows us to aim
for a global timeout on cgen/testing thm/defthm forms

remove all old references to ACL2::ALL. ACL2S:ALL is the new
TOP. Uppercase some quoted symbols. Move the default empty type alist
creation to the function make-dumb-type-alist in type.lisp and always
use that instead of pairlis

Added support for complex-rationals in number ranges.