Additional sanity check for books Makefiles
This patch adds an additional check to the books/Makefile to try to ensure that
whatever version of ACL2 you are using does, in fact, believe that you are in
the correct :dir :system.

This is mostly of use to folks who, for whatever reason, have multiple working
copies of the ACL2 repository and might occasionally, by accident, try to
certify books in one copy when the other ACL2 is in your path.