New timing tools
Oracle-time allows you to run a form and get its runtime and bytes allocated as
real ACL2 objects (logically, by reading the oracle).  I think this should be
sound.

Oracle-timelimit allows you to do the same, while also aborting the execution
of the form if it takes too long.  This is less sound (it could fail to
complete necessary stobj updates or similar), but may be practically useful.

  (Note: oracle-timelimit doesn't currently work well at the top level
   on failure)