Omitting statemen22.lisp on ACL2(r).
This book is slowly turning many of my builds red.  It takes a week
for them all to turn green again, so I'm going to disable it.  J's
patch will of course be very welcome if he makes one, but there's
not much reason to keep it enabled in the meantime.