Removed dependence on #+/#- for syscall numbers.
In the programmer-level mode of the x86isa model, the syscall numbers
used to be defined using defconst. E.g., the syscall number for the
read system call looked like:

(defconst *SYS_read*
  #+(and linux (not (or darwin bsd)))
  0
  #+(and (or darwin bsd) (not linux))
  #x2000003)

This was a problem sometimes; e.g., if the underlying Lisp *features*
had some variants of linux, darwin, or bsd, then the defconst would be
missing its body and the regression would fail.  Now, we define the
syscall numbers based on the value of the field os-info in the x86
state. The read syscall number is now defined as follows:

(define sys_read-idx ((x86))
  :enabled t
  (if (programmer-level-mode x86)
      (let ((os (os-info x86)))
        (cond
         ((equal os :linux)  0)
         ((equal os :darwin) #x2000003)
         (t -1)))
    nil))

A benefit of this is that unlike before, we can reason about linux
binaries while running our ACL2 session on a darwin machine and vice
versa. Of course, for the correct execution of binaries, os-info
should be initialized by the user appropriately, i.e., we can't
execute a linux binary on a darwin machine even if os-info is :linux.