Tolerate disabled subset in pick-a-point proofs
This adds a new rule, pick-a-point-subset-constraint-helper, to deal
with a weird constraint that arises during pick-a-point proofs of
subset, namely the proof that SUBSET is just a particular
implementation of ALL.

This constraint was normally discharged by just letting the definition
of subset open up, but (of course) that's not really desirable.
Historically, disabling subset's definition typically would screw up
pick-a-point proofs.  The new rule should help to avoid this and allow
pick-a-point proofs to be used even when subset is disabled.