Extend make-flag to generate templates
This extends the defthm-flag-xxx macros introduced by make-flag so
that they can print a template that you can copy and paste.  This may
be useful since there's a fair bit of syntax to remember and it's not
always easy to remember all the cases you need to cover, their flags,
the names of the formals, etc.

To get a template, just invoke the macro with no arguments to get a
template.  For instance,

    ACL2 !>(defthm-pseudo-termp)

Will print:

    Here's a template for using DEFTHM-PSEUDO-TERMP:

    (DEFTHM-PSEUDO-TERMP (DEFTHM THEOREM-FOR-PSEUDO-TERMP
				 (IMPLIES (AND HYP1 HYP2)
					  (PROP (PSEUDO-TERMP X)))
				 :FLAG TERM)
			 (DEFTHM THEOREM-FOR-PSEUDO-TERM-LISTP
				 (IMPLIES (AND HYP1 HYP2)
					  (PROP (PSEUDO-TERM-LISTP LST)))
				 :FLAG LIST))

    You'll probably want to adjust the names, hyps, and conclusion
    terms above.  Note also that you can use :skip, :rule-classes,
    etc.; for more information see :doc make-flag.