Fix xdoc preprocessor error handling bug
The error handling for bad @(`...`) forms was wrong and could lead
to a hard Lisp error by trying to concatenate a string with an MSG
style object.  Fix this (badly) to at least stringify the error
message instead of blowing up.

Fix a typo in the new make-flag documentation that provoked the
above error.