Tune rule and defrule via with-output
Preivously, RULE and DEFRULE would usually produce a lot of excessive, spurious
output.  This tries to use WITH-OUTPUT to cut down on this extra printing.

This patch is good enough to make the following successful thm/rules produce
almost identical output:

  (thm (equal (len (append x y)) (+ (len x) (len y)))
       :hints(("Goal" :in-theory (enable append))))

  ;; enable in :hints -- no encapsulate
  (rule (equal (len (append x y)) (+ (len x) (len y)))
        :hints(("Goal" :in-theory (enable append))))

  ;; top-level :enable -- causes an encapsulate
  (rule (equal (len (append x y)) (+ (len x) (len y)))
        :enable append)

The only difference is that the RULE forms don't print "Proof succeeded," which
may as well be considered a feature.  Successful DEFRULE versions of the above
are also now working without any extra printing.

The patch is also good enough to make the following failure cases very close:

  (thm (equal (len (append x y)) (cons (len x) (len y)))
       :hints(("Goal" :in-theory (enable append))))

  ;; enable in :hints -- no encapsulate
  (rule (equal (len (append x y)) (cons (len x) (len y)))
        :hints(("Goal" :in-theory (enable append))))

Unfortunately it's not quite good enough to get rid of all the extra printed
junk for this one:

  ;; top-level :enable -- causes an encapsulate
  (rule (equal (len (append x y)) (cons (len x) (len y)))
        :hints(("Goal" :in-theory (enable append))))

In particular we end up with a couple of extra lines like these, separated by a
lot of whitespace lines:

    ACL2 Error in ( PROGN (WITH-OUTPUT :STACK ...) ...):  PROGN failed!
    ACL2 Error in ( PROGN (WITH-OUTPUT :STACK ...) ...):  See :DOC failure.

So that's a bummer, but at least it's quite a lot better than before.

Errors in :prep-books and :prep-lemmas are still not handled very well, and
result in many unnecessary lines of printing past the error.  I don't know how
to fix this.  For instance, either of these still result in 30+ lines of
unfortunate extra printing:

   (defrule my-rule
     (equal (len (append x y)) (+ (len x) (len y)))
     :hints(("Goal" :in-theory (enable append)))
     :prep-books ((include-book "does-not-exist" :dir :system)))

   (defrule my-rule
     (equal (len (append x y)) (+ (len x) (len y)))
     :prep-lemmas ((local (defthm l0 (equal (cons x y) (+ x y))))))