Doc tweaks for some std/util macros
When we're using deflist/defalist/etc. to prove boilerplate theorems
about some already defined function, and when that function happens to
already be documented, then probably should keep the existing
documentation instead of overwriting it with the generic
deflist/defalist/etc. stuff.

So, try to do that by adjusting these macros to look for existing
documentation in the already-definedp case.