Scattered documentation fixes
Fix minor problems with the FTY autogenerated docs for defprods.  They
were, in the good case, showing the definition of the constructor
twice, and in bad case, trying to show a definition that didn't exist.

Rename the VL/VL2014 functions vl-moditem-alist and vl-portdecl-alist
to vl-make-moditem-alist and vl-make-portdecl-alist, to avoid doc
issues from trying to use the same name as an FTY type and also as a
function.

Adjust various other books to avoid other doc topic conflicts.