Add preferred topics to fixtypes
In the golden age of cutil and std/util, the closest thing to a type
was a recognizer, so recognizers like widget-p ended up being the
"main" topic for a type.

Now, with FTY, we tend to instead use the type name (e.g., widget) as
the main topic.  Pages like widget-p, widget-fix, widget-equiv, and
eat-widget being found underneath widget.  This is fine and seems to
be we want to standardize on going forward, so I'll call it the
"typename topic convention."

However, in the documentation produced by macros like defprod, where
we want to provide links to the types of each field, we were still
linking to recognizers, i.e., for vl-gateinst we were producing things
like this:

     type     --- vl-gatetype-p  ...
     name     --- maybe-stringp  ...
     atts     --- vl-atts-p      ...

Whereas really we should instead be linking to topics like vl-gatetype
and vl-atts.

This would be easy if everything followed the typename topic
convention, but various base types like nat/string/etc. don't follow
the convention, so in those cases we probably still want to link to
the recognizers, i.e., natp, stringp, etc.

To resolve this, we add a new TOPIC field to each fixtype, which can
be chosen at DEFFIXTYPE time, and which indicates the preferred topic
to link to when referring to this type.  The defaults topic is just
the type name (i.e., by default types are expected to follow the
typename topic convention), but for cases like natp/stringp we can
instead configure the TOPIC to point to the recognizer (or whatever
else is appropriate) instead of having to introduce new topics.

The idea is that macros like DEFPROD can now use this TOPIC field when
creating links to the type so that the documentation points to the
main page for each type, instead of just pointing to the (much less
descriptive) recognizer pages.

This patch tweaks DEFPROD to do this; it may be that other FTY macros
should also be similarly extended.