Improve with-output handling in defsection
Adjust DEFSECTION so that it doesn't explicitly turn on ERROR printing during
the submission of the events.  This helps to fix up some of the remaining
ugliness for RULE and DEFRULE printing.

Also, tweak DEFSECTION in various other ways, e.g., to make its error handling
more robust and make it a bit more strict in certain cases that came up during
testing the output stuff.