Fixing a VL bug.
In the event that the error is unknown, but size is still nil, the
fatal warning message would cause an error itself.  This was because
error argument 2 would be bound to nil, which isn't a valid input
for the tilde at format directive.

So, I've changed argument 2 to take on a somewhat artificial
error message in the case that err is nil.