Allow MONITOR to take symbols and other runic designators, and to warn on simple definition rules.  Added runic abbreviation :r.  Added book suggesting how Codewalker might allow one to skip or defer termination proofs.  Fixed a bug in remove-hyps.
Quoting :doc note-7-2:

  A new runic abbreviation has been added: :r, which abbreviates
  :rewrite. For example, you may use (:r app-assoc) in a theory
  expression to denote the [rune] (:rewrite app-assoc). See
  [theories].

  It is now possible to [monitor] rules by supplying a symbol or other
  runic designator other than a theory (see [theories]), thus
  relaxing the previous requirement that a [rune] is supplied. For
  example, the form :monitor nth t is legal, as is :monitor (:d nth)
  t, where before the corresponding command would necessarily have
  been :monitor (:definition nth) t. Thanks to Eric Smith for
  requesting this enhancement.

  Suitable [warnings] are now printed when attempting to [monitor]
  [simple] rules of class :[definition]. These had only been printed
  for simple rules of class :(tsee rewrite), because of potential
  inefficiency in computing for such warnings in the case of large
  [mutual-recursion] [events], but that problem has been addressed in
  the ACL2 source code using [fast-alists].

NOTE to anyone maintaining books/coi/super-ihs/symbols.lsp: in that
book are many lists of ACL2 symbols.  I suspect that some of those
lists are out of date.  In particular, this commit removes the
definitions of functions collect-non-backchain-subclass,
chk-acceptable-monitor and chk-acceptable-monitors from the ACL2
sources, and adds a few others.

The new book books/projects/codewalker/demo-fact-partial.lisp is a
variant of books/projects/codewalker/demo-fact.lisp that provides a
guide to how Codewalker might be modified so that termination proofs
can be avoided or delayed.

Fixed a bug in remove-hyps that occurs when no "proof steps" are
required in a proof (because the proof is so trivial).  Thanks to
Warren Hunt for reporting this bug.  Also: remove-hyps now runs
silently unless a new optional verbose argument is supplied.  See :doc
remove-hyps.