VL: dimension parsing changes, especially for modinsts
SystemVerilog allows modinst arrays like foo [3] (a, b, c), but
we weren't accepting dimensions like foo[3] instead of foo[3:0].
Fix that.  (SystemVerilog also allows multiple-dimension instance
arrays, but those are still unsupported.)

Also: we were parsing [] as a valid unpacked_dimension, but that
isn't permitted by the grammar.  Changed the parser to be more
restrictive.  This might break things.  If so, we probably need
to switch to use the variable_dimension parser instead of the
unpacked_dimension parser.