Three major updates to the x86isa books, described below:
----------------------------------------------------------------------
Modified the top-level definitions of memory accessor and updater functions.

Affected functions are: rm16, wm16, rm32, wm32, rm64, wm64, rm128,
wm128, write-canonical-address-to-memory.

The programmer-level mode behavior of these functions is the same as
it was before, but these functions have changed in the system-level
mode. We describe this change using the function rm16 as an example.

Previously, for (rm16 lin-addr r-w-x x86), we would figure out if both
lin-addr and (1+ lin-addr) fall in the same physical page. If so, we
would walk the page tables to translate the lin-addr to its
corresponding physical address phy-addr, and the physical address
corresponding to (1+ lin-addr) would simply be (1+ phy-addr).  This
helped avoid a page table walk for the translation of (1+ lin-addr)
and thus, increased the execution efficiency in the common case.

Now, we explicitly perform the translation for both lin-addr and (1+
lin-addr). This has definitely slowed down our simulation speed (but
only in the system-level mode). However, our optimizations were not
strictly correct. Because the second page walk was optimized away, our
functions essentially said that the second page walk never modifies
the x86 state, when this is not the case because a page walk may
modify the accessed and/or dirty bits in the translation-governing
entries of the paging data structures.

So, with this change, we lose on the execution efficiency in the
system-level mode, but gain accuracy.

Simulation speed in the System-Level Mode with identity page tables
(1GB pages):

  Old Model: Around 900K instructions/second
  New Model: Around 795K instructions/second

We can imagine "correctly" optimizing away the page walks by
performing the accessed/dirty bit updates without actually walking the
paging data structures, but that optimization will complicate
reasoning and thus, we haven't implemented it for now.

----------------------------------------------------------------------

Decreased the *initial-mem-array-pages* from 10 to 2. The memory
footprint of the model is smaller now --- the initial memory
requirement for mem-array went down from 1.2 GB to 256 MB. If more
initial memory is needed, simply change the value of
*initial-mem-array-pages* in portcullis/sharp-dot-constants.lisp ---
all the proofs should proceed smoothly.

Also sped up machine/x86-concrete-memory.lisp by reducing the number
of local include-books.

----------------------------------------------------------------------

Ported over FP instructions added by Cuong Kim Chau (thanks, Cuong!).

----------------------------------------------------------------------