Extend VL ~a fmt directive and fix up tag-reasoning ruleset
The tag-reasoning ruleset was defined in two places, leading to the
usual problems with it being smashed if books are included in
different orders; fix this by defining it only in std/util/da-base
(where tag is defined).

VL's ~a fmt directive didn't know about some of the new constructs, so
add them in and add a coverage proof to show that it properly
understands many kinds of structures.  This way, if we extend them in
the future, e.g., if we add additional kinds of statements or context
elements or whatever, then we'll remember to extend the ~a printing.