Startup banner now displays the git commit hash.  Eliminated redefinition errors involving enable.  Tweaked clausifier heuristics.
Quoting :doc note-7-2:

  For [history] commands and the function [disabledp], the notion of
  ``[disable]d'' is now computed inside the [break-rewrite] loop with
  respect to the [enable]d status at the current point of the current
  proof attempt, rather than with respect to the global state. Thanks
  to David Rager for suggesting this enhancement. Note that this
  change only affects [history] commands and [disabledp], not other
  utilities (for example [tau-status] and [guard-obligation]).x

  The startup banner now shows the git commit hash for development
  snapshots. Thanks to Bob Boyer for suggesting such a modification
  and to Jared Davis for bringing to our attention the appropriate
  git command for retrieving the hash.

  We eliminated errors for redefinition involving [hints] of the form
  (in-theory (enable ...)). Thanks to Sol Swords for sending a small
  example to exhibit the problem.

  We tweaked a part of the so-called ``clausify'' process for doing
  Boolean simplification towards producing subgoals of a given goal,
  namely, the so-called ``subsumption-replacement-loop''
  (specifically, source function find-subsumer-replacement). Thanks
  to Jared Davis for helpful discussions that led us to make this
  change.

The first of the items above replaces an earlier one, which had been
recently added, for a utility with-brr-ens.  That utility has been
deleted.  Thanks to David for convincing me to make some further
change so that :pr etc. automatically show the enabled status with
respect to the current proof point.

The last of the items above stemmed from a discussion with Jared,
which started when he sent a code modification for function
find-subsumer-replacement that reduced a particular proof's time from
4.77 seconds to 1.73 seconds.  That modification avoided a certain
computation when a clause is too long.  In his experiment he took a
length to be "too long" when it exceeded 10; we did something else.
I'd be very interested to know if anyone comes up with a better change
to find-subsumer-replacement; perhaps then I'd make the new part of
the code attachable, or perhaps better yet, make further changes.
Jared has made a good point (in an email ot me) about how it could be
challenging to arrange for user-defined attachments to work well with
future prover mods.