The specification of some legacy SSE/SSE2 instructions is now taken from rtl/rel11.
In addition to FP rounding and other basic functions, we rely on
rtl/rel11 for the specifications of addition, multiplication,
subtraction, division, and square root operations.  Home-grown FP
specifications include those of FP minimum, maximum, comparisons, and
converts (they aren't yet present in rtl/rel11). Thanks a lot to Cuong
Kim Chau for writing and testing the x86isa FP specifications!

The x86isa books leverage years of work done on the rtl books, which
are very well vetted. This decision also enables the developers of the
x86isa books to focus on their current research goal --- developing
capabilities to reason about system software --- without spending too
much time on specifying and testing yet more x86 FP
instructions. Supporting more instructions in the x86isa interpreter
will almost always be a task that the developers will perform as
needed, and we are hopeful that switching to rtl/rel11 for our FP
specifications will allow others to execute and reason about their FP
application programs on the x86isa interpreter.

It would be simply wonderful if the specification functions in
rtl/rel11 had guards and type declarations (and also no consing ---
e.g., formatp), both for ease of comprehension and for execution
efficiency. But that's a task and gripe for another day.