Forall and exists in defun-sk must be in the "ACL2" package.  Defun redundancy now depends on :normalize.
  The quantifier [forall] or [exists] used in the body of a [defun-sk]
  call must now be in the "ACL2" package. Many users will not notice
  this change, since the symbols forall and exists in the "ACL2"
  package are now in the list [*ACL2-exports*]; see the paragraph
  just above. Thanks to Alessandro Coglio for a remark leading to
  this change.

  Definitions are no longer redundant that have different values
  specified for their [xargs] keyword, :normalize. (See
  [redundant-events].) Although we are not aware of a need to make
  this change, this change eliminates a possibile unsoundness.

Also, made a minor improvement in an error message for :use hints to
mention :termination-theorem :termination-theorem!, and
:guard-theorem.

Deleted some obsolete comments.

Improved :doc xargs to clarify the effect of declaring (xargs :guard
t).  Thanks to Cuong for a question leading to this clarification.

Added missing symbols GETPROPC, GTHM, and TTHM to *acl2-exports* (not
caught in regression during certification attempt for
books/misc/check-acl2-exports.lisp, because of how I did it in
stages).